+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-- Keyboard input handling system
+- Keyboard input handling system (event dispatch and querying functionality)
 
 ---
 ## 0.0.1-indev-1 - 2021-11-28 - The init commit